download: Add support to an user defined progress bar
This commit just introduces a new argument to the download function: `bar`.
The argument let the person who is developing code define his own progress
bar using whatever he wants to. It also requires the pattern of passing
the current size of the downloaded chunk. This callable function respects the
tqdm API.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>